requirements.txt
flask>=3.0
openai>=1.0
python-dotenv>=1.0
ddgs>=2.0
requests>=2.31
beautifulsoup4>=4.12
pypdf>=5.0
chromadb>=0.5
sentence-transformers>=3.0
torch>=2.3
torchvision>=0.18
transformers>=4.49
accelerate>=1.2
bitsandbytes>=0.43
qwen-vl-utils>=0.0.8
Pillow>=10.0
